FIX IO for large files

+++ include/rawop.h
@@ -22,9 +22,9 @@ namespace io {
                        uint64_t fromPosition,
                        uint8_t* buffer,
                        std::streamsize numBytes);
-    void writeFirstBytes(std::string fileName, uint8_t* buffer, int numBytes);
-    void writeBytesFrom(std::string fileName, uint64_t fromPosition, uint8_t* buffer, int numBytes);
-    void appendBytes(std::string fileName, uint8_t* buffer, int numBytes);
+    void writeFirstBytes(std::string fileName, uint8_t* buffer, uint64_t numBytes);
+    void writeBytesFrom(std::string fileName, uint64_t fromPosition, uint8_t* buffer, uint64_t numBytes);
+    void appendBytes(std::string fileName, uint8_t* buffer, uint64_t numBytes);
     void createEmptyFile(std::string fileName, uint64_t numBytes, bool overwrite);
     bool fileExists(std::string fileName);
     long getFileSize(std::string filename);
